NAT-PMP port forwarding support, temporarily merge codes to resolve conflicts.
