+++ .gitmodules
@@ -2,3 +2,7 @@
     path = thirdparty/miniupnp
     url = https://github.com/miniupnp/miniupnp.git
     ignore = dirty
+[submodule "thirdparty/libnatpmp"]
+    path = thirdparty/libnatpmp
+    url = https://github.com/miniupnp/libnatpmp.git
+    ignore = dirty
