+++ CMakeLists.txt
@@ -6,7 +6,7 @@ SET(CMAKE_VERBOSE_MAKEFILE ON)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 # N2n release information
-set(N2N_VERSION "2.9.0")
+set(N2N_VERSION "3.1.0")
 set(N2N_OSNAME ${CMAKE_SYSTEM_NAME})
 execute_process(
         COMMAND git status
@@ -87,6 +87,7 @@ if(N2N_OPTION_USE_MINIUPNP)
   ADD_DEFINITIONS("-DN2N_HAVE_MINIUPNP")
   include_directories(${THIRD_PARTY_DIR}/miniupnp/miniupnpc/include)
   include_directories(${PROJECT_BINARY_DIR}/lib_miniupnpc)
+  include_directories(${THIRD_PARTY_DIR}/libnatpmp)
 endif(N2N_OPTION_USE_MINIUPNP)
 
 if(NOT DEFINED CMAKE_BUILD_TYPE)
@@ -154,7 +155,7 @@ add_library(n2n STATIC
         src/sn_selection.c
         src/auth.c
         src/curve25519.c
-        src/n2n_upnpc.c)
+        src/n2n_port_mapping.c)
 
 
 if(N2N_OPTION_USE_PTHREAD)
@@ -173,7 +174,9 @@ endif(N2N_OPTION_USE_ZSTD)
 if(N2N_OPTION_USE_MINIUPNP)
   add_subdirectory(${THIRD_PARTY_DIR}/miniupnp/miniupnpc lib_miniupnpc)
   link_directories(${PROJECT_BINARY_DIR}/lib_miniupnpc)
-  target_link_libraries(n2n libminiupnpc-static)
+  add_subdirectory(${THIRD_PARTY_DIR}/libnatpmp libnatpmp)
+  link_directories(${PROJECT_BINARY_DIR}/libnatpmp)
+  target_link_libraries(n2n libminiupnpc-static natpmp)
 endif(N2N_OPTION_USE_MINIUPNP)
 
 if(DEFINED WIN32)
